editor: a column is a long, not a file offset
edit_draw_this_line() kept the visual column in an off_t and passed its
address to edit_get_line_offset(), whose out parameter is a long *. Where
off_t is long long, as it is on macOS, that is a pointer type mismatch and
the callee would write eight bytes through a pointer to four or write in the
wrong layout; on x86_64 Linux the two types coincide and nothing complains,
which is why it went unnoticed.

The column is a long everywhere else: in the parameter, in the counters
inside edit_get_line_offset(), in start_col_real which consumes it here, and
in the same local in edit_move_to_prev_col() and edit_save_block_to_clip().
So the declaration follows suit, the one place that read a byte distance into
it says so with a cast, and the cast that only existed to narrow it back to a
long is gone.

+++ src/editor/editdraw.c
@@ -505,7 +505,7 @@ edit_draw_this_line (WEdit *edit, off_t b, long row, long start_col, long end_co
     line_s *p = line;
     off_t q;
     off_t first_visible;
-    off_t col;
+    long col;
     long start_col_real;
     int abn_style;
     int book_mark = 0;
@@ -540,7 +540,7 @@ edit_draw_this_line (WEdit *edit, off_t b, long row, long start_col, long end_co
         const long target_col = MAX (0, start_col - edit->start_col);
 
         q = MIN (edit->buffer.size, b + (off_t) target_col);
-        col = q - b;
+        col = (long) (q - b);
     }
     else
     {
@@ -623,7 +623,7 @@ edit_draw_this_line (WEdit *edit, off_t b, long row, long start_col, long end_co
                     else
                     {
                         // col is this cell's absolute visual column
-                        const long x = (long) col;
+                        const long x = col;
                         long cl;
 
                         cl = MIN (edit->column1, edit->column2);
